server/requirements.txt
alabaster==1.0.0
asgiref==3.8.1
attrs==23.2.0
babel==2.16.0
black==24.8.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
cryptography==42.0.5
Django==5.1.1
django-stubs==5.1.0
django-stubs-ext==5.1.0
djangorestframework==3.15.2
djangorestframework-stubs==3.15.1
docutils==0.21.2
exceptiongroup==1.2.2
flake8==7.1.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
iniconfig==2.0.0
isort==5.13.2
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
keyring==23.13.1
lxml==4.9.4
MarkupSafe==2.1.5
mccabe==0.7.0
more-itertools==10.2.0
mypy==1.11.2
mypy-extensions==1.0.0
openconnect-sso==0.8.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
pyotp==2.9.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
PyQt6-WebEngine==6.7.0
PyQt6-WebEngine-Qt6==6.7.0
PySocks==1.7.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
pyxdg==0.28
requests==2.32.3
SecretStorage==3.3.3
setuptools==69.5.1
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
structlog==24.1.0
toml==0.10.2
tomli==2.0.1
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
zipp==3.20.2
